requirements.txt
websockets==10.4
Pillow==9.4.0
numpy==1.24.2
opencv-python==4.7.0.72
matplotlib==3.7.1
pywinauto==0.6.8
